Moved and renamed some classes
