+++ app/src/main/java/com/sylvainmrzd/avivrealestate/feature_ads/presentation/ad_detail/viewmodel/AdDetailViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.viewModelScope
-import com.sylvainmrzd.avivrealestate.feature_ads.data.repository.AvivRealEstateRepository
+import com.sylvainmrzd.avivrealestate.feature_ads.domain.repository.AvivRealEstateRepository
 import com.sylvainmrzd.avivrealestate.feature_ads.domain.model.Ad
 import com.sylvainmrzd.avivrealestate.feature_ads.presentation.BaseViewModel
 import com.sylvainmrzd.avivrealestate.feature_ads.presentation.util.AdElements
